requirements.txt
django>=3.1,<3.2
dj-database-url==0.5.0
django-storage-url==0.5.0
whitenoise==5.2.0
boto3==1.14.49

psycopg2==2.8.5

uwsgi==2.0.19.1
django-translation-manager==1.1.1
python-gettext==4.0
easy-thumbnails
django-taggit
django-tinymce